fix(aoe/clean.sh): Don't use grep to find files
Changes:
- Replace grepping in listing outputs with find
- Replace acting on files using xargs with the -exec function of find
- Add indentions to improve readability

The output ls is mainly designed to be human readable, therefore parsing it's
output is fragile. Also grep is a tool for pattern matching, that doesn't know
about filenames and operates on a "line" base (at least by default). This can
cause false positives and therefore errors. Since this script already uses
find and therefore requires this tool to be present, we might as well use it.
This also allows to ditch xargs, since it can be substituted with finds own
functionality (-exec).